requirements.txt
attrs==24.3.0
black==24.10.0
cachetools==5.5.1
cffi==1.17.1
chardet==5.2.0
click==8.1.8
colorama==0.4.6
coverage==7.6.10
debugpy==1.8.12
distlib==0.3.9
eccodes==2.39.2
filelock==3.17.0
findlibs==0.0.5
flake8==7.1.1
gitdb==4.0.12
GitPython==3.1.44
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
pyproject-api==1.9.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-flakes==4.0.5
smmap==5.0.2
tox==4.24.1
virtualenv==20.29.1
